restructure folder organization and improve logic

+++ src/services/order_service.js
@@ -1,16 +1,12 @@
 import db from "../models";
-import { Nearest } from "../utils/NearestStock";
-export const addOrder = async (data) => {
+import { calculateDistance } from "../helper/calculateDistance";
+import { sendEmailOrder } from "../helper/sendEmail";
+import { CalculateTotalPrice } from "../helper/discountHelper";
+export const createOrder = async (data) => {
 	try {
-		const NOW = new Date();
-		const dateOnly = NOW.toISOString().split("T")[0];
-		let total_price = 0;
-		total_price = Number.parseFloat(total_price);
-		const discount = await db.Discount_code.findOne({
-			where: {
-				code: data.discount_code,
-			},
-		});
+		const productVariantData = [];
+
+		// Lấy dữ liệu từ giỏ hàng
 		const CartData = await db.Cart.findOne({
 			where: {
 				user_id: data.user_id,
@@ -20,91 +16,31 @@ export const addOrder = async (data) => {
 				attributes: ["product_variant_id", "quantity"],
 			},
 		});
-		for (const e of CartData.Cart_items) {
-			console.log(e);
-		}
 
-		if (data.discount_code) {
-			if (!discount) {
-				throw { status: 400, message: "Voucher code does not exist" };
-			}
-			const checkUseDiscount = await db.Order.findOne({
-				where: {
-					discount_id: discount.id,
-				},
-			});
-			if (checkUseDiscount) {
-				throw { status: 400, message: "Voucher code already used" };
-			}
-			if (dateOnly <= discount.expiry) {
-				throw { status: 400, message: "Voucher code has expired" };
-			}
-		}
+		const foundEmail = await db.User.findOne({
+			where: {
+				id: data.user_id,
+			},
+		});
+		const totalInfo = await CalculateTotalPrice(
+			data.user_id,
+			data.discount_code,
+			CartData.Cart_items,
+		);
+		console.log(totalInfo);
 
-		for (const q of CartData.Cart_items) {
-			const productData = await db.Product_variant.findByPk(
-				q.product_variant_id,
-			);
-
-			if (discount) {
-				if (discount.discount_type === "percentage_discount") {
-					if (discount.target_type === "all") {
-						productData.price =
-							productData.price - (discount.value * productData.price) / 100;
-					} else if (
-						discount.target_type === "product" &&
-						discount.target_id === q.product_variant_id
-					) {
-						productData.price =
-							productData.price - (discount.value * productData.price) / 100;
-					} else if (
-						discount.target_type === "category" &&
-						discount.target_id === productData.category_id
-					) {
-						productData.price =
-							productData.price - (discount.value * productData.price) / 100;
-					} else if (
-						discount.target_type === "brand" &&
-						discount.target_id === productData.brand_id
-					) {
-						productData.price =
-							productData.price - (discount.value * productData.price) / 100;
-					}
-				}
-				if (discount.discount_type === "fixed_amount_discount") {
-					if (discount.target_type === "all") {
-						productData.price = productData.price - discount.value;
-					} else if (
-						discount.target_type === "product" &&
-						discount.target_id === q.product_variant_id
-					) {
-						productData.price = productData.price - discount.value;
-					} else if (
-						discount.target_type === "category" &&
-						discount.target_id === productData.category_id
-					) {
-						productData.price = productData.price - discount.value;
-					} else if (
-						discount.target_type === "brand" &&
-						discount.target_id === productData.brand_id
-					) {
-						productData.price = productData.price - discount.value;
-					}
-					total_price += q.quantity * productData.price;
-				}
-			} else {
-				total_price += q.quantity * productData.price;
-			}
-		}
+		// Tạo dữ liệu order
 		const response = await db.Order.create({
 			user_id: data.user_id,
-			total_amount: total_price + (total_price * 0.015 + 15000),
-			discount_code_id: discount ? discount.id : null,
+			total_amount: totalInfo.totalPrice,
+			discount_code_id: totalInfo.foundCode ? totalInfo.foundCode.id : null,
 		});
+		// Tạo các order item dựa trên dữ liệu của giỏ hàng
 		for (const q of CartData.Cart_items) {
 			const productData = await db.Product_variant.findByPk(
 				q.product_variant_id,
 			);
+			productVariantData.push(productData);
 			await db.Order_item.create({
 				order_id: response.id,
 				product_variant_id: q.product_variant_id,
@@ -112,19 +48,26 @@ export const addOrder = async (data) => {
 				price: productData.price * q.quantity,
 			});
 		}
-		await db.Shipping.create({
-			order_id: response.id,
-			address_id: data.shipping.address_id,
-			type: data.shipping.type,
-			status: "pending",
-			shipfee: total_price * 0.015 + 15000,
+
+		// Lưu các thông tin liên quan đến ship.
+		// if (!data.shipping) {
+		// 	throw { status: 400, message: "Vui lòng cập nhật địa chỉ nhận hàng!" };
+		// }
+		// await db.Shipping.create({
+		// 	order_id: response.id,
+		// 	address_id: data.shipping.address_id,
+		// 	type: data.shipping.type,
+		// 	status: "pending",
+		// 	shipfee: total_price * 0.015 + 15000,
+		// });
+		// Xóa giỏ hàng khi đã checkout thành công
+		await db.Cart.destroy({
+			where: {
+				user_id: data.user_id,
+			},
 		});
 
-		// await db.Cart.destroy({
-		// 	where: {
-		// 		user_id: data.user_id,
-		// 	},
-		// });
+		sendEmailOrder(foundEmail.email, productVariantData);
 		return {
 			data: response,
 		};
@@ -135,11 +78,18 @@ export const addOrder = async (data) => {
 };
 export const deleteOrder = async (param) => {
 	try {
+		const foundOrder = await db.Order.findByPk(param);
+		if (!foundOrder) {
+			throw {
+				status: 404,
+				message: "Đơn đặt hàng không tồn tại, xóa không thành công!",
+			};
+		}
 		await db.Order.destroy({
 			where: { id: param },
 		});
 
-		return { message: "Delete Successfully" };
+		return { message: "Xóa thành công!" };
 	} catch (error) {
 		console.error(error);
 		throw new Error("Database Error");
@@ -163,79 +113,89 @@ export const updateOrder = async (data, param) => {
 				},
 			],
 		});
-		// Kiểm tra request status được gửi
-		if (orderData.status !== data.status) {
-			await db.Order.update(
-				{
-					status: data.status,
-				},
-				{
-					where: { id: param },
-				},
-			);
-			// Cập nhật lại tồn kho
-			if (data.status === "shipping") {
-				for (const e of orderData.Order_items) {
-					let nereastStock = null;
-					let minDistance = Number.POSITIVE_INFINITY;
-					const findProduct = await db.Product_variant.findByPk(
-						e.product_variant_id,
-						{
-							include: {
-								model: db.Inventory,
-								attributes: ["id", "quantity", "location"],
-							},
+		// Kiểm tra tránh cùng một trạng thái cập nhật nhiều lần
+		if (orderData.status === data.status) {
+			throw {
+				status: 400,
+				message: `Trạng thái đang là:${orderData.status} `,
+			};
+		}
+		await db.Order.update(
+			{
+				status: data.status,
+			},
+			{
+				where: { id: param },
+			},
+		);
+		// Cập nhật lại tồn kho khi status shipping
+		if (data.status === "shipping") {
+			for (const e of orderData.Order_items) {
+				let nereastStock = null;
+				let minDistance = Number.POSITIVE_INFINITY;
+				const findProduct = await db.Product_variant.findByPk(
+					e.product_variant_id,
+					{
+						include: {
+							model: db.Inventory,
+							attributes: ["id", "quantity", "location"],
 						},
+					},
+				);
+
+				for (const s of findProduct.Inventories) {
+					const inventory = await db.Inventory.findByPk(s.id);
+					// Tính khoảng cách từ địa chỉ giao hàng đến vị trí của kho
+					const distance = await calculateDistance(
+						orderData.Shipping.Address.address_line_1,
+						inventory.location,
 					);
-
-					for (const s of findProduct.Inventories) {
-						const findStock = await db.Inventory.findByPk(s.id);
-						// Kiểm tra khoảng cách giữa kho và địa chỉ nhận hàng
-
-						console.log(`log address: ${orderData.Shipping.Address}`);
-
-						const distance = await Nearest(
-							orderData.Shipping.Address.address_line_1,
-							findStock.location,
-						);
-						if (distance < minDistance && s.quantity > e.quantity) {
-							minDistance = distance;
-							nereastStock = s;
-						}
+					// Kiểm tra xem kho này có gần hơn kho hiện tại và còn đủ hàng không
+					if (distance < minDistance && s.quantity > e.quantity) {
+						minDistance = distance;
+						nereastStock = s;
 					}
-					if (!nereastStock) {
-						throw { status: 400, message: "Out of Inventories" };
-					}
-					await db.Inventory.update(
-						{ quantity: nereastStock.quantity - e.quantity },
-						{
-							where: {
-								id: nereastStock.id,
-							},
-						},
-					);
-					console.log(
-						`Cập nhật tồn kho thành công tại kho ${nereastStock.id} cho sản phẩm ${e.product_variant_id}`,
-					);
 				}
+				if (!nereastStock) {
+					throw {
+						status: 400,
+						message: "Kho hết hàng! không thể đổi trạng thái!",
+					};
+				}
+				await db.Inventory.update(
+					{ quantity: nereastStock.quantity - e.quantity },
+					{
+						where: {
+							id: nereastStock.id,
+						},
+					},
+				);
+				console.log(
+					`Cập nhật tồn kho thành công tại kho ${nereastStock.id} cho sản phẩm ${e.product_variant_id}`,
+				);
 			}
-
-			return { message: "Update Successfully", data: orderData };
 		}
 
-		throw { status: 400, message: `Order status ${orderData.status} already` };
+		return { message: "Cập nhật trạng thái thành công", data: orderData };
 	} catch (error) {
 		console.error(error);
 		throw error;
 	}
 };
-export const getOrder = async (user) => {
+
+export const getAllOrder = async (user) => {
 	try {
 		const response = await db.Order.findAll({
 			include: [
 				{
 					model: db.Order_item,
-					attributes: ["id", "product_variant_id", "quantity", "price"],
+					attributes: [
+						"id",
+						"product_variant_id",
+						"quantity",
+						"price",
+						"status",
+					],
 				},
 				{
 					model: db.Shipping,
@@ -253,6 +213,6 @@ export const getOrder = async (user) => {
 		}
 	} catch (error) {
 		console.error(error);
-		throw new Error("Database Error");
+		throw error;
 	}
 };
